refactor: like Entity, Repository 위치 변경
